+++ app/components/app.js
@@ -1,293 +1,25 @@
-import angular from 'angular';
-
-import find from 'lodash/find';
-
-import {
-  realms as REALMS,
-  enRealmNames as EN_REALM_NAMES
-} from '../config';
-import template from './app.html';
-
 export const NAME = 'app';
 
-const AUTO_REQUERY_DIFF = 86400 * 1000;  // auto re-query
+const template = `
+<div class="r-app">
+  <div class="r-app__content" ng-transclude="content"></div>
+  <div class="r-app__footer" ng-transclude="footer">
+    <div class="md-caption c-footer-info">
+      <a href="https://github.com/pc035860/hi-armory.tw" rel="noopener" target="_blank"><i class="fa fa-github-alt"></i>pc035860/hi-armory.tw</a>
+    </div>
+  </div>
+</div>
+`;
 
-const DEFAULT_REGION = 'tw';
-
-class AppCtrl {
-  static $inject = [
-    '$log', '$scope', '$state', 'wowProfile', 'parseProfile', 'ga', '$location',
-    'charIndex', '$mdMedia', 'closeKeyboard', '$window', 'realmName', 'wclId',
-    '$timeout'
-  ];
-
-  __deps;
-  _realmSearchList = {};
-  REALMS;
-  profile;
-  wclId;
-  region;
-  realm;
-  character;
-  autocompleteItem = {
-    realm: null,
-    character: null
-  };
-  reloading;
-  pp;  // parsedProfile
-  pd;  // profile.data
-
-  /* @ngInject */
-  constructor(
-    $log, $scope, $state, wowProfile, parseProfile, ga, $location,
-    charIndex, $mdMedia, closeKeyboard, $window, realmName, wclId,
-    $timeout
-  ) {
-    $scope.ga = ga;
-
-    this.region = DEFAULT_REGION;
-    this.fields = 'items,progression,talents';
-    this.REALMS = REALMS;
-
-    const rn = realmName(this.region);
-
-    this.__deps = { $log, wowProfile, ga, charIndex, $window, rn, $scope, $timeout };
-
-    $scope.$watch(() => $state.params, (val, oldVal) => {
-      const region = this.region;
-      const { realm: rawRealm, character } = val;
-
-      $log.debug('val', rawRealm, character);
-
-      const requireNewProfile = !angular.equals(val, oldVal) || !this.profile;
-
-      if (this.profile && requireNewProfile) {
-        this.profile.$destroy();
-      }
-
-      if (region && rawRealm && character) {
-        const realm = rn.isEn(rawRealm) ? rn.toLocaled(rawRealm) : rawRealm;
-        Object.assign(this, {
-          region,
-          realm,
-          character
-        });
-
-        if (requireNewProfile) {
-          closeKeyboard();
-
-          this.profile = wowProfile.getFirebaseObject({
-            region,
-            realm,
-            character,
-            fields: this.fields
-          });
-
-          this.profile.$loaded()
-          .then((obj) => {
-            if (!obj.data && obj.status !== 'not found') {
-              // no data, trigger first-time query
-              this.query({ enqueue: true });
-            }
-            else if ((+new Date()) - obj.dataUpdatedAt > AUTO_REQUERY_DIFF) {
-              // auto re-query
-              this.query({ enqueue: true });
-            }
-          });
-
-          this.wclId = wclId.getFirebaseObject({ region, realm, character });
-
-          this.wclId.$loaded()
-          .then((obj) => {
-            if (!obj.$value) {
-              wclId.enqueue({
-                region,
-                realm,
-                character
-              });
-            }
-          });
-        }
-      }
-      else {
-        this.profile = null;
-      }
-    });
-
-    $scope.$watch(() => this.profile, (val, oldVal) => {
-      if (val && val.data && typeof val.data !== 'string') {
-        this.pp = parseProfile(val.data);  // parsedProfile
-        this.pd = val.data;
-      }
-      else {
-        this.pp = null;
-        this.pd = null;
-      }
-
-      if (val && val.status && oldVal && oldVal.status) {
-        if (val.status !== oldVal.status) {
-          // 只是介面上的 reloading 意味
-          this.reloading = false;
-        }
-      }
-
-      if (val && val.status && val.status === 'ready') {
-        charIndex.addHistory(`${this.region}-${this.realm}-${this.character}`);
-      }
-    }, true);
-
-    $scope.$watch(() => $mdMedia('xs'), (xs) => {
-      this.mqXS = xs;
-    });
-
-    $scope.$on('$stateChangeSuccess', () => {
-      ga.pageview($location.path());
-    });
-  }
-
-  /**
-   * Query
-   * @param  {object} options   { bool: enqueue  }
-   */
-  query(options_) {
-    const { $scope, wowProfile } = this.__deps;
-
-    if ($scope.armoryForm.$invalid) {
-      return;
-    }
-
-    const { region, realm, character } = this;
-
-    if (!region || !realm || !character) {
-      return;
-    }
-
-    if (
-      this.profile && this.profile.$resolved &&
-      this.profile.status !== 'ready' && this.profile.status !== 'not found' &&
-      this.profile.$value !== null) {
-      return;
-    }
-
-    const options = {
-      ...{ enqueue: false },
-      ...(options_ || {})
-    };
-
-    const srefParams = {
-      region: 'tw',
-      realm,
-      character
-    };
-
-    if (options.enqueue) {
-      wowProfile.enqueue({
-        ...srefParams,
-        fields: this.fields
-      });
-    }
-
-    wowProfile.goToState(srefParams);
-  }
-
-  reload() {
-    this.query({ enqueue: true });
-    this.reloading = true;
-  }
-
-  _updateSelectedRealmItem(region, realm) {
-    const [local, preEn] = find(this.REALMS[this.region], v => v[0] === realm);
-    this.autocompleteItem.realm = AppCtrl.getRealmItem(local, preEn);
-  }
-
-  realmSearch(str) {
-    const realms = this.REALMS[this.region];
-    // 建立內部搜尋用的 map
-    if (!this._realmSearchList[this.region]) {
-      this._realmSearchList[this.region] = realms.map(([local, preEn]) => {
-        const en = EN_REALM_NAMES[preEn].split(' ').join('');
-        return `${local}${en}`;
-      });
-    }
-
-    const re = new RegExp(str, 'i');
-    return realms
-    .filter((v, i) => re.test(this._realmSearchList[this.region][i]))
-    .map(([local, preEn]) => {
-      return AppCtrl.getRealmItem(local, preEn);
-    });
-  }
-
-  indexSearch(str) {
-    const { charIndex } = this.__deps;
-    return charIndex.search(str);
-  }
-
-  handleSelectedRealmItemChange(item) {
-    if (!item) {
-      return;
-    }
-    const realm = item.local;
-    Object.assign(this, { realm });
-  }
-
-  handleSelectedCharacterItemChange(item) {
-    if (!item) {
-      return;
-    }
-    const { $timeout } = this.__deps;
-    const { region, realm } = item;
-    Object.assign(this, { region, realm });
-
-    this._updateSelectedRealmItem(region, realm);
-
-    $timeout(() => {
-      this.query();
-    });
-  }
-
-  outbound(target) {
-    const { $window, rn, ga } = this.__deps;
-
-    const enRealmName = rn.toEn(this.realm);
-    const realm = encodeURIComponent(this.realm);
-    const character = encodeURIComponent(this.character);
-
-    let link;
-
-    switch (target) {
-      case 'armory':
-        link = `http://${this.region}.battle.net/wow/zh/character/${realm}/${character}/advanced`;
-        break;
-      case 'wowprogress':
-        link = `https://www.wowprogress.com/character/${this.region}/${enRealmName}/${character}`;
-        break;
-      case 'warcraftlogs':
-        link = `https://www.warcraftlogs.com/rankings/character/${this.wclId.$value}/latest/`;
-        break;
-      default:
-        break;
-    }
-
-    if (link) {
-      $window.open(link);
-    }
-
-    ga.event('Outbound Link', 'click', target);
-  }
-
-  static getRealmItem(local, preEn) {
-    const en = EN_REALM_NAMES[preEn];
-    return {
-      local, en
-    };
-  }
-}
+class AppCtrl { }
 
 const component = {
+  transclude: {
+    content: 'appContent',
+    footer: '?appFooter'
+  },
   controller: AppCtrl,
-  controllerAs: '$app',
-  template
+  template,
 };
 
 export default function configure(ngModule) {
