+++ app/sass/_components.scss
@@ -394,3 +394,35 @@
 		}
 	}
 }
+
+.c-app-heading {
+	position: relative;
+
+	&__switch-link {
+		$layout-padding: 8px;
+
+		@include bigger-hotzone(before, 100px, 36px);
+
+		position: absolute;
+		right: $layout-padding;
+		bottom: $layout-padding + 10px;
+
+		font-size: 14px;
+		line-height: 1;
+		color: #ff6f00 !important;
+
+		@media #{$mq-xs} {
+			bottom: -5px;
+		}
+	}
+}
+
+.c-footer-info {
+	.fa {
+		position: relative;
+		top: -1px;
+		margin-right: 3px;
+		vertical-align: middle;
+		font-size: 18px;
+	}
+}
